requirements.txt
asgi-redis==1.4.1
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.1
Automat==0.6.0
certifi==2017.4.17
channels==1.1.5
chardet==3.0.4
codecov==2.0.9
constantly==15.1.0
coverage==4.4.1
daphne==1.3.0
Django==1.11.2
djangorestframework==3.6.3
flake8==3.3.0
gevent==1.2.2
greenlet==0.4.12
hyperlink==17.2.0
idna==2.5
incremental==17.5.0
Jinja2==2.9.6
jsonfield==2.0.1
MarkupSafe==1.0
mccabe==0.6.1
model-mommy==1.3.2
msgpack-python==0.4.8
pycodestyle==2.3.1
pyflakes==1.5.0
pytz==2017.2
redis==2.10.5
requests==2.18.1
six==1.10.0
Twisted==17.5.0
txaio==2.8.0
urllib3==1.21.1
websocket==0.2.1
websocket-client==0.42.1
zope.interface==4.4.2
